Working on calculating data, now can evaluate simple expressions like 2+2, 5*2, 7/7 etc

+++ backend/models/expression.go
@@ -4,16 +4,16 @@ import "time"
 
 // Expression выражение
 type Expression struct {
-	ID         int     `json:"id"`
-	Expression string  `json:"expression"`
-	Status     string  `json:"status"`
-	Result     *int    `json:"result,omitempty"`
-	CreatedAt  string  `json:"created_at"`
-	FinishedAt *string `json:"finished_at,omitempty"`
+	ID         int      `json:"id"`
+	Expression string   `json:"expression"`
+	Status     string   `json:"status"`
+	Result     *float64 `json:"result,omitempty"`
+	CreatedAt  string   `json:"created_at"`
+	FinishedAt *string  `json:"finished_at,omitempty"`
 }
 
 // ChangeData Изменение данных выражения
-func (e *Expression) ChangeData(status string, result int) {
+func (e *Expression) ChangeData(status string, result float64) {
 	e.Status = status
 	e.Result = &result
 	finish := time.Now().Format("02-01-2006 15:04:05")
